+++ bioblend/galaxy/objects/wrappers.py
@@ -1025,19 +1025,21 @@ class HistoryDatasetAssociation(Dataset):
         self.__init__(res, self.container, gi=self.gi)  # type: ignore[misc]
         return self
 
-    def delete(self, purge: bool = False) -> None:
+    def delete(self, purge: bool = False, wait: bool = False) -> None:
         """
         Delete this history dataset.
 
         :type purge: bool
         :param purge: if ``True``, also purge (permanently delete) the dataset
 
+        :param wait: Whether to wait for the dataset to be purged.
+
         .. note::
           The ``purge`` option works only if the Galaxy instance has the
           ``allow_user_dataset_purge`` option set to ``true`` in the
           ``config/galaxy.yml`` configuration file.
         """
-        self.gi.gi.histories.delete_dataset(self.container.id, self.id, purge=purge)
+        self.gi.gi.histories.delete_dataset(self.container.id, self.id, purge=purge, wait=wait)
         self.container.refresh()
         self.refresh()